+++ app/handlers/apiv1/post.go
@@ -1,0 +1,19 @@
+package apiv1
+
+import "github.com/getfider/fider/app/pkg/web"
+
+// SearchPosts return existing posts based on search criteria
+func SearchPosts() web.HandlerFunc {
+	return func(c web.Context) error {
+		posts, err := c.Services().Posts.Search(
+			c.QueryParam("q"),
+			c.QueryParam("f"),
+			c.QueryParam("l"),
+			c.QueryParamAsArray("t"),
+		)
+		if err != nil {
+			return c.Failure(err)
+		}
+		return c.Ok(posts)
+	}
+}
