+++ app/handlers/post.go
@@ -50,22 +50,6 @@ func Index() web.HandlerFunc {
 	}
 }
 
-// SearchPosts return existing posts based on search criteria
-func SearchPosts() web.HandlerFunc {
-	return func(c web.Context) error {
-		posts, err := c.Services().Posts.Search(
-			c.QueryParam("q"),
-			c.QueryParam("f"),
-			c.QueryParam("l"),
-			c.QueryParamAsArray("t"),
-		)
-		if err != nil {
-			return c.Failure(err)
-		}
-		return c.Ok(posts)
-	}
-}
-
 // CreatePost creates a new post on current tenant
 func CreatePost() web.HandlerFunc {
 	return func(c web.Context) error {
